refactor(app_frame): split AppFrame into mixin package
Move widgets/frames/app_frame.py into a package preserving the existing
AppEventHandlers/SideboardGuideHandlers/CardTablePanelHandler mixin
chain. frame.py holds the __init__, all _build_* sub-builders, and
launch_app; properties.py exposes _t, _set_status, the deck-result
delegation @property accessors, and _has_deck_loaded; handlers.py owns
the toolbar settings menu, _apply_* setting handlers, session restore,
window persistence, timer scheduling, fetch_archetypes, and deck
rendering helpers.
