+++ widgets/frames/app_frame/__init__.py
@@ -1,0 +1,7 @@
+"""Main application frame package."""
+
+from __future__ import annotations
+
+from widgets.frames.app_frame.frame import AppFrame, launch_app
+
+__all__ = ["AppFrame", "launch_app"]
